Add editing and deleting routes

+++ Application/Routes/Commands/DeleteRoute/DeleteRouteCommandHandler.cs
@@ -1,20 +1,15 @@
-﻿using System;
-using System.Linq;
+﻿using Application.Common.Exceptions;
 using Application.Common.Interfaces;
+using Domain.Entities;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
+using System;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Application.Common.Exceptions;
-using Domain.Entities;
-using Microsoft.EntityFrameworkCore;
 
-namespace Application.Routes.Commands
+namespace Application.Routes.Commands.DeleteRoute
 {
-    public class DeleteRouteCommand : IRequest
-    {
-        public int Id { get; set; }
-    }
-
     public class DeleteRouteCommandHandler : IRequestHandler<DeleteRouteCommand>
     {
         private readonly IApplicationDbContext _context;
@@ -29,6 +24,7 @@ namespace Application.Routes.Commands
             var entity = await _context.Routes
                 .Include(r => r.StartingStation)
                 .Include(r => r.FinalStation)
+                .Include(r => r.Train)
                 .Where(r => r.Id == request.Id)
                 .SingleOrDefaultAsync(cancellationToken);
 
@@ -40,6 +36,12 @@ namespace Application.Routes.Commands
             if (!entity.IsOnHold)
             {
                 throw new InvalidOperationException("A route cannot be deleted unless it's on hold.");
+            }
+
+            if (entity.Train is not null)
+            {
+                throw new InvalidOperationException("A route cannot be deleted if there are trains assigned to it." +
+                                                    " First, mark the route as suspended and unassign the trains to the route.");
             }
 
             _context.Routes.Remove(entity);
